Add functional link and route to movie details screen.
Co-Authored By: Jarrett Kong <jarrettkong@me.com>
