+++ src/components/MovieDetails/MovieDetails.jsx
@@ -11,9 +11,8 @@ class MovieDetails extends Component {
 		this.setState({ loading: true }, async () => {
 			try {
 				const res = await fetch(`${REACT_APP_BASE_URL}/movie/${this.props.id}?api_key=${REACT_APP_API_KEY}`);
-				const json = await res.json();
-				console.log(json.results);
-				this.setState({ loading: false });
+				const details = await res.json();
+				this.setState({ details, loading: false });
 			} catch (err) {
 				console.log(err);
 			}
@@ -22,7 +21,12 @@ class MovieDetails extends Component {
 
 	render() {
 		const display = this.state.loading ? <h1>Loading...</h1> : <h1>Done.</h1>;
-		return <section>{display}</section>;
+		return (
+			<section>
+				<h1>hi</h1>
+				{display}
+			</section>
+		);
 	}
 }
