+++ src/components/Poster/Poster.jsx
@@ -4,12 +4,10 @@ import './_Poster.scss';
 
 const Poster = props => {
 	return (
-		<Link to={`/movies/title/${props.id}`} key={props.id}>
 		<article className="Poster">
 			<img src={props.poster} alt={`${props.title} poster`} className="Poster-poster" />
 			<h3 className="Poster-title">{props.title}</h3>
 		</article>
-		</Link>
 	);
 };
