+++ src/containers/MovieContainer/MovieContainer.jsx
@@ -1,10 +1,17 @@
 import React from 'react';
 import Poster from '../../components/Poster/Poster';
 import { connect } from 'react-redux';
+import { Link } from 'react-router-dom';
 import './_MovieContainer.scss';
 
 export const MovieContainer = props => {
-	const posters = props.movies.slice(0, 6).map(movie => <Poster key={movie.id} {...movie} />);
+  const posters = props.movies.slice(0, 6).map(movie => {
+    return (
+      <Link to={`/title/${movie.id}`} key={movie.id}>
+        <Poster key={movie.id} {...movie} />
+      </Link>
+    );
+  });
 	return (
 		<section className={`MovieContainer ${props.section}`}>
 			<h2 className="MovieContainer-section-title">{`${props.section.toUpperCase()}`}</h2>
